refactor RholangCLI
* keep the IO stuff in the RholangCLI
* move all the functions related to interpreting Rholang code to Interpreter
